requirements.txt
gradio
fastapi
uvicorn
python-multipart
langchain
langchain-huggingface
langchain-chroma
langchain-openai
pydantic
python-dotenv
pytesseract
pillow
pymupdf